+++ wsa_client.h
@@ -21,13 +21,13 @@
 ///////////////////////////////////////////////////////////////////////////////
 // Global Functions
 ///////////////////////////////////////////////////////////////////////////////
-int32_t wsa_list_ips(char **ip_list);
+int16_t wsa_list_ips(char **ip_list);
 u_long wsa_verify_addr(const char *sock_addr);
-int32_t wsa_get_host_info(char *name);
-int32_t wsa_start_client(const char *wsa_addr, SOCKET *cmd_sock, 
+int16_t wsa_get_host_info(char *name);
+int16_t wsa_start_client(const char *wsa_addr, SOCKET *cmd_sock, 
 						 SOCKET *data_sock);
-int32_t wsa_close_client(SOCKET cmd_sock, SOCKET data_sock);
-int32_t wsa_sock_send(SOCKET out_sock, char *out_str, int32_t len);
-int32_t wsa_sock_recv(SOCKET in_sock, char *rx_buf_ptr, uint32_t time_out);
+int16_t wsa_close_client(SOCKET cmd_sock, SOCKET data_sock);
+int16_t wsa_sock_send(SOCKET out_sock, char *out_str, int32_t len);
+int64_t wsa_sock_recv(SOCKET in_sock, char *rx_buf_ptr, uint32_t time_out);
 
 #endif
